takeMS that returns initial segments
In the case of trees, we want it to return the list of paths of depth n.
takeMMS' and takeMMS'' return many spurious results because they recombine
heads and tails in all possible way.

To avoid this: get heads and tail out using unwrapMS
so a head is recombined only with its original tail.