initMMS'': correction and simplification using mapOutMS
mapOutMS allows us to specify a function directly on head and tail
then it is applied inside the monadic action automatically.

This is similar to transformMS for functors
It allows a uniform application that doesn't proliferate the monad actions

initMMS'' was not working correctly for trees:
the null test applied to the "tail" with was a list of tails
it was empty only if there were no tails on any branch.
Instead we want to check if each branch is null.