Explanation of guarded-by-recursion and def of functorial mapping
I changed the biblio file to monster_biblio.bib because
it conflicted with the global bib file that I have on my laptop.

I defined functorial mapping in a more informal-categorical way.
I think this is easier to read and to work with.
We should use this notation for proofs rather than Haskell.