Definition of applicative methods from monad
I didn't manage to prove the third applicative law in general
but I can prove it if M is a monad.
So I introduce here the translation from monad to applicative operators.
Then I will prove a couple of lemmas relating the two notations
and finally the third (interchange) law.