Very large cut-down of content in the monad section
I've included (I believe) just enough exposition to explain the rationale behind the counterexample and definition of join.

Things I've cut:
 - Large diagrams graphically explaining the join function
 - Formal proof that the left identity monad law doesn't hold

This section can only be included if the code is made public really, since the explanations are quite a bit terser and up to the reader to fill in the details. Especially since we don't give a definition of `run` in the paper